made order controller

+++ backend/src/main/java/com/Ecommerce/backend/entity/OrderItem.java
@@ -2,7 +2,12 @@ package com.Ecommerce.backend.entity;
 
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -15,7 +20,7 @@ public class OrderItem {
     private long id;
 
     @ManyToOne
-    @JoinColumn(name ="order_id")
+    @JoinColumn(name = "order_id")
     private Order order;
 
     @ManyToOne
@@ -24,5 +29,5 @@ public class OrderItem {
 
     private int Quantity;
 
-    private double price;
+    private BigDecimal price;
 }
